+++ src/tool/cppwinrt/cppwinrt/code_writers.h
@@ -67,6 +67,28 @@ namespace xlang
         w.write(format);
     }
 
+    static void write_parent_depends(writer& w, cache const& c, std::string_view const& type_namespace)
+    {
+        auto pos = type_namespace.rfind('.');
+
+        if (pos == std::string::npos)
+        {
+            return;
+        }
+
+        auto parent = type_namespace.substr(0, pos);
+        auto found = c.namespaces().find(parent);
+
+        if (found != c.namespaces().end() && has_projected_types(found->second))
+        {
+            w.write_root_include(parent);
+        }
+        else
+        {
+            write_parent_depends(w, c, parent);
+        }
+    }
+
     static void write_pch(writer& w)
     {
         auto format = R"(#include "%"
