+++ src/tool/cppwinrt/cppwinrt/helpers.h
@@ -548,4 +548,14 @@ namespace xlang
 
         return false;
     }
+
+    static bool has_projected_types(cache::namespace_members const& members)
+    {
+        return
+            !members.interfaces.empty() ||
+            !members.classes.empty() ||
+            !members.enums.empty() ||
+            !members.structs.empty() ||
+            !members.delegates.empty();
+    }
 }
